requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==1.5.3
langchain==0.3.13
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-community==0.3.13
faiss-cpu==1.7.4
pydantic==2.10.4
numpy==1.25.0
python-dotenv==1.0.1
